app/requirements.txt
requests==2.32.3
PyYAML==6.0.2
fastapi==0.112.2
uvicorn==0.30.6
pytest==8.3.2
pytest-mock==3.14.0
prometheus-client==0.20.0


